Revert "320 bug fix add mock data"
