--- smssync/src/internal/assets/json/logs/log.json
@@ -1,4 +1,0 @@
-{
-  "id": 2,
-  "message": "09-09 13:25 There is no active data connection"
-}
