--- smssync/src/internal/assets/json/logs/logs.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": 1,
-    "message": "09-09 13:25 HTTP Entity format is [URLEncoded]"
-  },
-  {
-    "id": 2,
-    "message": "09-09 13:25 There is no active data connection"
-  },
-  {
-    "id": 3,
-    "message": "17:34 13:25 Sync started"
-  },
-  {
-    "id": 4,
-    "message": "09-09 13:25 The device has shut-down at 17:59"
-  },
-  {
-    "id": 5,
-    "message": "09-47 13:25 Sync started"
-  },
-  {
-    "id": 6,
-    "message": "09-35 13:25 HTTP Entity format is [URLEncoded]"
-  }
-]
