--- smssync/src/internal/assets/json/messages/message.json
@@ -1,13 +1,0 @@
-{
-  "to": "+000-000-0000",
-  "message": "the message goes here",
-  "uuid": "aada21b0-0615-4957-bcb3",
-  "timestamp": "1435153499",
-  "sent_result_code": 0,
-  "sent_result_message": "SMSSync Message Sent",
-  "delivered_result_code": 0,
-  "delivered_result_message": "SMS Delivered",
-  "retries": 0,
-  "status": "failed",
-  "type": "pending"
-}
