--- smssync/src/internal/assets/json/messages/task_messages.json
@@ -1,106 +1,0 @@
-[
-  {
-    "to": "+000-000-0000",
-    "message": "the message goes here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "failed",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "our message is here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "sent",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "bank u",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "failed",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "Fishing boats",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "sent",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "the message goes here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "sent",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "the message goes here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "Not sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "Not Delivered",
-    "retries": 0,
-    "status": "failed",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "the message goes here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "sent",
-    "type": "task"
-  },
-  {
-    "to": "+000-000-0000",
-    "message": "the message goes here",
-    "uuid": "aada21b0-0615-4957-bcb3",
-    "timestamp": "1435153499",
-    "sent_result_code": 0,
-    "sent_result_message": "SMSSync Message Sent",
-    "delivered_result_code": 0,
-    "delivered_result_message": "SMS Delivered",
-    "retries": 0,
-    "status": "sent",
-    "type": "task"
-  }
-]
