--- smssync/src/internal/java/org/addhen/smssync/data/repository/DataHelper.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2010 - 2015 Ushahidi Inc
- * All rights reserved
- * Contact: team@ushahidi.com
- * Website: http://www.ushahidi.com
- * GNU Lesser General Public License Usage
- * This file may be used under the terms of the GNU Lesser
- * General Public License version 3 as published by the Free Software
- * Foundation and appearing in the file LICENSE.LGPL included in the
- * packaging of this file. Please review the following information to
- * ensure the GNU Lesser General Public License version 3 requirements
- * will be met: http://www.gnu.org/licenses/lgpl.html.
- *
- * If you have questions regarding the use of this file, please contact
- * Ushahidi developers at team@ushahidi.com.
- */
-
-package org.addhen.smssync.data.repository;
-
-import android.content.Context;
-
-import java.io.IOException;
-import java.io.InputStream;
-
-/**
- * @author Ushahidi Team <team@ushahidi.com>
- */
-public final class DataHelper {
-
-    private DataHelper() {
-    }
-
-    public static String loadJSONFromAsset(Context context, final String jsonFileName) {
-        String json;
-        try {
-            InputStream is = context.getAssets().open("json/" + jsonFileName);
-            int size = is.available();
-            byte[] buffer = new byte[size];
-            is.read(buffer);
-            is.close();
-            json = new String(buffer, "UTF-8");
-        } catch (IOException ex) {
-            ex.printStackTrace();
-            return "";
-        }
-        return json;
-    }
-}
